move validity date logic from DateRestrictingNode to ValidityDateNode
