+++ backend/src/main/java/com/bakdata/conquery/apiv1/query/concept/specific/CQDateRestriction.java
@@ -1,10 +1,8 @@
 package com.bakdata.conquery.apiv1.query.concept.specific;
 
 import java.time.LocalDate;
-import java.util.ArrayDeque;
 import java.util.Collections;
 import java.util.List;
-import java.util.Queue;
 import java.util.function.Consumer;
 
 import javax.validation.Valid;
@@ -23,8 +21,6 @@ import com.bakdata.conquery.models.query.Visitable;
 import com.bakdata.conquery.models.query.queryplan.ConceptQueryPlan;
 import com.bakdata.conquery.models.query.queryplan.QPNode;
 import com.bakdata.conquery.models.query.queryplan.specific.DateRestrictingNode;
-import com.bakdata.conquery.models.query.queryplan.specific.NegatingNode;
-import com.bakdata.conquery.models.query.queryplan.specific.ValidityDateNode;
 import com.bakdata.conquery.models.query.resultinfo.ResultInfo;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -43,31 +39,12 @@ public class CQDateRestriction extends CQElement {
 
 	@Override
 	public QPNode createQueryPlan(QueryPlanContext context, ConceptQueryPlan plan) {
-		QPNode childAgg = child.createQueryPlan(context.withDateRestriction(CDateRange.of(dateRange)), plan);
+		QPNode childQueryPlan = child.createQueryPlan(context.withDateRestriction(CDateRange.of(dateRange)), plan);
 
-		//insert behind every ValidityDateNode
-		Queue<QPNode> openList = new ArrayDeque<>();
-
-		openList.add(childAgg);
-
-		while (!openList.isEmpty()) {
-			QPNode current = openList.poll();
-			if (current instanceof ValidityDateNode validityDateNode) {
-
-				validityDateNode.setChild(new DateRestrictingNode(
-						CDateSet.create(Collections.singleton(CDateRange.of(dateRange))),
-						validityDateNode.getChild()
-				));
-			}
-			else if (current instanceof NegatingNode) {
-				//we can't push date restrictions past negations
-			}
-			else {
-				openList.addAll(current.getChildren());
-			}
-		}
-
-		return childAgg;
+		return new DateRestrictingNode(
+				CDateSet.create(Collections.singleton(CDateRange.of(dateRange))),
+				childQueryPlan
+		);
 	}
 
 	@Override
