+++ backend/src/main/java/com/bakdata/conquery/models/query/queryplan/specific/DateRestrictingNode.java
@@ -1,14 +1,8 @@
 package com.bakdata.conquery.models.query.queryplan.specific;
 
-import java.util.Map;
-import java.util.Objects;
-
 import com.bakdata.conquery.models.common.CDateSet;
-import com.bakdata.conquery.models.common.daterange.CDateRange;
-import com.bakdata.conquery.models.datasets.Column;
 import com.bakdata.conquery.models.datasets.Table;
 import com.bakdata.conquery.models.events.Bucket;
-import com.bakdata.conquery.models.events.CBlock;
 import com.bakdata.conquery.models.query.QueryExecutionContext;
 import com.bakdata.conquery.models.query.queryplan.QPChainNode;
 import com.bakdata.conquery.models.query.queryplan.QPNode;
@@ -22,8 +16,6 @@ import lombok.ToString;
 public class DateRestrictingNode extends QPChainNode {
 
 	protected final CDateSet restriction;
-	protected Column validityDateColumn;
-	protected Map<Bucket, CBlock> preCurrentRow;
 
 	public DateRestrictingNode(CDateSet restriction, QPNode child) {
 		super(child);
@@ -42,35 +34,10 @@ public class DateRestrictingNode extends QPChainNode {
 			ctx = ctx.withDateRestriction(dateRestriction);
 		}
 		super.nextTable(ctx, currentTable);
-
-
-		preCurrentRow = ctx.getBucketManager().getEntityCBlocksForConnector(getEntity(), context.getConnector());
-		validityDateColumn = context.getValidityDateColumn();
-
-		if (validityDateColumn != null && !validityDateColumn.getType().isDateCompatible()) {
-			throw new IllegalStateException("The validityDateColumn " + validityDateColumn + " is not a DATE TYPE");
-		}
-	}
-
-	@Override
-	public boolean isOfInterest(Bucket bucket) {
-		final CBlock cBlock = Objects.requireNonNull(preCurrentRow.get(bucket));
-
-		if (validityDateColumn == null) {
-			// If there is no validity date set for a concept there is nothing to restrict
-			return super.isOfInterest(bucket);
-		}
-
-		final CDateRange range = cBlock.getEntityDateRange(entity.getId());
-
-		return restriction.intersects(range) && super.isOfInterest(bucket);
 	}
 
 	@Override
 	public void acceptEvent(Bucket bucket, int event) {
-		if (validityDateColumn != null && !bucket.eventIsContainedIn(event, validityDateColumn, restriction)) {
-			return;
-		}
 		getChild().acceptEvent(bucket, event);
 	}
