+++ backend/src/main/java/com/bakdata/conquery/models/query/queryplan/specific/ExternalNode.java
@@ -80,7 +80,7 @@ public class ExternalNode<T> extends QPNode {
 
 	@Override
 	public boolean isContained() {
-		return contained != null;
+		return contained != null && !dateUnion.isEmpty();
 	}
 
 	@Override
